requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2023.2
sqlparse==0.4.3
